Tried fixing the Driver class yet again, and I am pretty sure that it works this time! The Driver class now usually creates about 150 rooms per second unless the rooms are requested to be very big (it could be made faster by scaling the size of the rooms down before requesting them to be placed).
I fixed some access modifiers in the Room class.

I made it so that the method place of the class Room properly checks for conflicting room positions, taking into account the sizes of the other rooms. The method place also doesn't replace the width and height properties of all the Room objects passed in as arguments annymore.