Started tinkering with the GUI, so it's messy right now.
Fixed Room's constructor so that it doesn't set its coordinates via arguments, and instead lets Driver do it with a setter. This however makes the coordinates not able to be final which kind of sucks since I don't like the idea of the room's coordinates changing after its construction. Rooms now consider connections to itself as nonexistent (null) now.

I fixed how Driver accessed its arrays, and now it's about 10 times faster depending on the size of the rooms being placed. This can still be made faster by scaling the rooms sizes while coming up with placements if we want to, but it might not be necessary.